+++ src/container/ProjectList.tsx
@@ -1,9 +1,9 @@
 import { VFC } from 'react';
-import useFindAllProjects from '../hooks/use-find-all-projects';
 import ProjectList from '../components/ProjectList';
+import { useFindAllProject } from '../service/project/project';
 
 const EnhancedProjectList: VFC = () => {
-  const { projects, isLoading } = useFindAllProjects();
+  const { projects, isLoading } = useFindAllProject();
 
   // eslint-disable-next-line react/react-in-jsx-scope
   return <ProjectList {...{ projects, isLoading }} />;
